+++ src/Support/Str.php
@@ -15,4 +15,22 @@ class Str extends \Illuminate\Support\Str
     {
         return preg_match($pattern, $string) != false;
     }
+
+    /**
+     * Determines the full url based on the available optional query parameters. Filters out null values of the
+     * parameters.
+     *
+     * @param string $url
+     * @param array $parameters
+     * @return string
+     */
+    public static function optionalQueryParameters(string $url, array $parameters): string
+    {
+        $parameters = array_filter($parameters);
+        if (count($parameters) === 0) {
+            return $url;
+        }
+
+        return $url . '?' . http_build_query($parameters);
+    }
 }